add pluginCheck.js to tensorflow plugin
Purpose is for when a plugin is started but crashes right away due to some kind of incompatibility with the users hardware. Some users may not notice this and let it run anyway. This will suspend the process if it stops and starts again within 3 seconds of the last startup.